Updated postman collection and outdated env
Docker now creates to 3001 and quite a few other adjustments still needed to be updated. This will get the postman tests to be 21 pass as well instead. Postman collection is as up-to-date till swagger gets updated as well or further clarification and I can make adjustments to swagger as well.